# perf: allow skipping public key verification