--- a/lib/hdkey.js
+++ b/lib/hdkey.js
@@ -41,6 +41,13 @@ Object.defineProperty(HDKey.prototype, 'privateKey', {
   }
 })
 
+function setPublicKey (hdkey, publicKey) {
+  hdkey._publicKey = Buffer.from(publicKey)
+  hdkey._identifier = hash160(publicKey)
+  hdkey._fingerprint = hdkey._identifier.slice(0, 4).readUInt32BE(0)
+  hdkey._privateKey = null
+}
+
 Object.defineProperty(HDKey.prototype, 'publicKey', {
   get: function () {
     return this._publicKey
@@ -48,16 +55,9 @@ Object.defineProperty(HDKey.prototype, 'publicKey', {
   set: function (value) {
     assert(value.length === 33 || value.length === 65, 'Public key must be 33 or 65 bytes.')
     assert(secp256k1.publicKeyVerify(value) === true, 'Invalid public key')
-
-    if (value.length === 65) {
-      // force compressed point (performs public key verification)
-      this._publicKey = Buffer.from(secp256k1.publicKeyConvert(value, true))
-    } else {
-      this._publicKey = Buffer.from(value)
-    }
-    this._identifier = hash160(this.publicKey)
-    this._fingerprint = this._identifier.slice(0, 4).readUInt32BE(0)
-    this._privateKey = null
+    // force compressed point (performs public key verification)
+    const publicKey = (value.length === 65) ? secp256k1.publicKeyConvert(value, true) : value
+    setPublicKey(this, publicKey)
   }
 })
 
@@ -194,7 +194,7 @@ HDKey.fromMasterSeed = function (seedBuffer, versions) {
   return hdkey
 }
 
-HDKey.fromExtendedKey = function (base58key, versions) {
+HDKey.fromExtendedKey = function (base58key, versions, skipVerification) {
   // => version(4) || depth(1) || fingerprint(4) || index(4) || chain(32) || key(33)
   versions = versions || BITCOIN_VERSIONS
   var hdkey = new HDKey(versions)
@@ -215,7 +215,11 @@ HDKey.fromExtendedKey = function (base58key, versions) {
     hdkey.privateKey = key.slice(1) // cut off first 0x0 byte
   } else {
     assert(version === versions.public, 'Version mismatch: version does not match public')
-    hdkey.publicKey = key
+    if (skipVerification) {
+      setPublicKey(hdkey, key)
+    } else {
+      hdkey.publicKey = key
+    }
   }
 
   return hdkey