--- a/lib/hdkey.js
+++ b/lib/hdkey.js
@@ -197,6 +197,7 @@ HDKey.fromMasterSeed = function (seedBuffer, versions) {
 HDKey.fromExtendedKey = function (base58key, versions, skipVerification) {
   // => version(4) || depth(1) || fingerprint(4) || index(4) || chain(32) || key(33)
   versions = versions || BITCOIN_VERSIONS
+  skipVerification = skipVerification || false
   var hdkey = new HDKey(versions)
 
   var keyBuffer = bs58check.decode(base58key)