The public key verification is rather slower in a pure JS mode. This PR introduces an argument  `skipVerification` to `fromExtendedKey()` to allow skipping the verification.$

builds on #52 

I did some quick benchmarks for thisPR. 

elliptic with BN.js: 8x performance increase with `skipVerification` set to true.
```
HDKey.fromExtendedKey(key) x 4,712 ops/sec ±4.69% (75 runs sampled)
HDKey.fromExtendedKey(key, null, true) x 39,627 ops/sec ±4.94% (71 runs sampled)
Fastest is HDKey.fromExtendedKey(key, null, true)
```

This allows `fromExtendedKey` to run just as quick as the native performance:
```
HDKey.fromExtendedKey(key) x 30,868 ops/sec ±4.71% (77 runs sampled)
HDKey.fromExtendedKey(key, null, true) x 38,105 ops/sec ±5.65% (67 runs sampled)
Fastest is HDKey.fromExtendedKey(key, null, true)
```