@RyanZim 2023-01-06
https://github.com/cryptocoinjs/hdkey/pull/52 is merged; please rebase.