@kewde 2023-01-14
@RyanZim I think there's already a unit test for it.